refactor common elastic code and bumpi index
